# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api de estudiantes prueba en servidor LINUX",
        "description": "Listado de URL de la api de estudiantes",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/api/usuarios": {
            "get": {
                "tags": [
                    "Maestro_usuarios"
                ],
                "summary": "Obtiene el listado de usuarios",
                "description": "Retorna una lista de todos los usuarios. Requiere un token de autenticación Bearer.",
                "operationId": "3d324e728a1a5172f62dfb8ebdb1b9ea",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "John Doe"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "johndoe@example.com"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "3112112111"
                                            },
                                            "language": {
                                                "type": "string",
                                                "example": "Spanish"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": "2024-02-23T00:09:16.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "example": "2024-02-23T12:33:45.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 401
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/{login}": {
            "get": {
                "tags": [
                    "Maestro_usuarios"
                ],
                "summary": "Obtiene un usuario por login",
                "description": "Retorna un usuario basado en su login. Requiere un token de autenticación Bearer.",
                "operationId": "9a2fff7c6bacba7b0658041bb0fe798d",
                "parameters": [
                    {
                        "name": "login",
                        "in": "path",
                        "description": "Login del usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "johndoe@example.com"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "3112112111"
                                                },
                                                "language": {
                                                    "type": "string",
                                                    "example": "Spanish"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-02-23T00:09:16.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-02-23T12:33:45.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No data found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No data found"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 401
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "apiUserAuth": {
                "type": "apiKey",
                "name": "x-api-user",
                "in": "header"
            },
            "apiKeyAuth": {
                "type": "apiKey",
                "name": "x-api-key",
                "in": "header"
            }
        }
    }
}